[DQ_CoppeliaSimInterfaceZMQ{Experimental}][Murilo] Splitting methods between the two classes. Cleaning up repeated methods in the Experimental class.
